# A Bunch Of Optimizations That Helped With A Large Image Network