--- a/include/GafferScene/DeleteSceneContextVariables.h
+++ b/include/GafferScene/DeleteSceneContextVariables.h
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (c) 2013, Image Engine Design Inc. All rights reserved.
+//  Copyright (c) 2017, Image Engine Design Inc. All rights reserved.
 //
 //  Redistribution and use in source and binary forms, with or without
 //  modification, are permitted provided that the following conditions are
@@ -34,29 +34,18 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#include "boost/python.hpp"
+#ifndef GAFFER_DELETESCENECONTEXTVARIABLES_H
+#define GAFFER_DELETESCENECONTEXTVARIABLES_H
 
-#include "GafferBindings/DependencyNodeBinding.h"
+#include "Gaffer/DeleteContextVariables.h"
+#include "GafferScene/SceneProcessor.h"
 
-#include "GafferScene/SceneContextProcessor.h"
-#include "GafferScene/SceneContextVariables.h"
-#include "GafferScene/DeleteSceneContextVariables.h"
-#include "GafferScene/SceneSwitch.h"
-#include "GafferScene/SceneTimeWarp.h"
-#include "GafferScene/SceneLoop.h"
-
-#include "GafferSceneBindings/MixinBinding.h"
-
-using namespace GafferScene;
-
-void GafferSceneBindings::bindMixin()
+namespace GafferScene
 {
 
-	GafferBindings::DependencyNodeClass<SceneContextProcessor>();
-	GafferBindings::DependencyNodeClass<SceneTimeWarp>();
-	GafferBindings::DependencyNodeClass<SceneContextVariables>();
-	GafferBindings::DependencyNodeClass<DeleteSceneContextVariables>();
-	GafferBindings::DependencyNodeClass<SceneSwitch>();
-	GafferBindings::DependencyNodeClass<SceneLoop>();
+typedef Gaffer::DeleteContextVariables<SceneProcessor> DeleteSceneContextVariables;
+IE_CORE_DECLAREPTR( DeleteSceneContextVariables )
 
-}
+} // namespace GafferScene
+
+#endif // GAFFER_DELETESCENECONTEXTVARIABLES_H