… python/GafferSceneUI/DeleteSceneContextVariablesUI.py
… python/GafferTest/DeleteContextVariablesTest.py
… python/GafferUI/DeleteContextVariablesUI.py
… python/GafferUI/__init__.py
… src/Gaffer/Context.cpp
… src/Gaffer/DeleteContextVariablesNode.cpp
… src/GafferBindings/ContextBinding.cpp
… src/GafferImage/DeleteImageContextVariables.cpp
… src/GafferImage/Grade.cpp
… src/GafferImage/ImageLoop.cpp
… src/GafferImage/ImageTimeWarp.cpp
… src/GafferImage/OpenImageIOReader.cpp
… src/GafferImage/Shuffle.cpp
… src/GafferImageBindings/MixinBinding.cpp
… src/GafferModule/GafferModule.cpp
… src/GafferScene/DeleteSceneContextVariables.cpp
… src/GafferScene/SceneLoop.cpp
… src/GafferScene/SceneTimeWarp.cpp
… startup/gui/menus.py