I haven't done final profiling of the end result of all this stuff yet, because there is one big fix that requires a new node that I'm working on, which I'll probably present as a separate PR.

But I didn't want to wait for that to have you take a look at this - the majority of these fixes seem fairly positive to me.  Let me know if there are specific things here that you want to see detailed performance analysis of.