@johnhaddon 2017-06-06
As you say, the thrust of this all makes sense, but I do think the extra complexities need to be justified with some basic benchmarking. Maybe you could provide some before/after numbers?