@danieldresser-ie 2017-06-09
Comments addressed.  Just need to find something reasonable to test on now.  Still tempted to temporarily rebase this onto CollectImages for profiling, so that I can test on a production case.