@johnhaddon 2017-06-09
> Just need to find something reasonable to test on now. Still tempted to temporarily rebase this onto CollectImages for profiling, so that I can test on a production case.

Testing the two together on the denoiser sounds fine to me. Maybe that plus a quick test on a more run-of-the-mill output profile?