@danieldresser-ie 2017-06-10
I've now rebased this on top of CollectImages, since my practical performance test scene required it as a prereq, and it seems like it might actually be easier to merge?  I can redo in the opposite order if you would prefer.

Summarizing Performance Results:

On the large network that originally concerned me:

|         | Time(s) | Memory(G) |
| ------ | ------ | ---------- |
| Initial | 202 | 19.5 |
| Using CollectImages | 72 | 8.5 |
| Using DeleteContextVariables | 46 | 6.1 |
| Using GlobalContext in reader | 43 | 5.5 |

The remaining commits ( using GlobalScope in Shuffle, Grade, TimeWarp, and Loop, and not leaking the Loop index out the main loop input ) did not have any noticeable impact on performance, either positive or negative.

Nevertheless, I strongly feel they are worthwhile.  Testing on just one layer instead of a full run of an executable the writes several layers, with the stats on, produced these numbers:

With just the obviously useful commits ( omitting the last 5 ):
```
  Max resident size    4326.953M
  Loading             9.368s (wall), 7.560s (CPU)
  Image generation    6.160s (wall), 68.700s (CPU)
  Total sum of time spent in hash and compute processes    : 96.938145 seconds
  Total time spent in hash processes                       : 36.742132 seconds
  Total time spent in compute processes                    : 60.196013 seconds
  Total time spent per hash process                        : 0.000012 seconds
  Total time spent per compute process                     : 0.000138 seconds
  Total number of hash processes                           : 3021308
  Total number of compute processes                        : 437126
  Total number of hash processes per compute process       : 6.911755
```

With the last 5 commits included:
```
  Max resident size    4324.793M
  Loading             8.437s (wall), 7.460s (CPU)
  Image generation    5.798s (wall), 68.530s (CPU)
  Total sum of time spent in hash and compute processes    : 91.434778 seconds
  Total time spent in hash processes                       : 37.547862 seconds
  Total time spent in compute processes                    : 53.886916 seconds
  Total time spent per hash process                        : 0.000013 seconds
  Total time spent per compute process                     : 0.000126 seconds
  Total number of hash processes                           : 2948773
  Total number of compute processes                        : 427129
  Total number of hash processes per compute process       : 6.903706
```

So, very little difference, though there does seem to be a couple percent improvement in speed that is above the noise level ( I'm seeing a fair bit of random variation in these tests, so it's a bit hard to tell ).  The small reduction in total number of hash processes does seem consistent.

The key factor with these hash fixes is that the hashing overhead is only noticeable when a bunch of things multiply together and blow up.  I think these fixes reduce the chances of that happening, and make it clearer to hunt down what's going wrong if we do hit a performance issue.

On a production slapcomp:
Before all changes:
```
 Max resident size    3542.555M

Performance :

  Loading             9.598s (wall), 8.010s (CPU)
  Image generation    8.548s (wall), 68.730s (CPU)

PerformanceMonitor Summary :

  Total sum of time spent in hash and compute processes    : 104.844986 seconds
  Total time spent in hash processes                       : 27.273429 seconds
  Total time spent in compute processes                    : 77.571557 seconds
  Total time spent per hash process                        : 0.000007 seconds
  Total time spent per compute process                     : 0.000248 seconds
  Total number of hash processes                           : 3896608
  Total number of compute processes                        : 313371
  Total number of hash processes per compute process       : 12.434488
```

After all changes:
```
  Max resident size    3462.746M

Performance :

  Loading             10.531s (wall), 8.410s (CPU)
  Image generation    7.860s (wall), 61.560s (CPU)

PerformanceMonitor Summary :

  Total sum of time spent in hash and compute processes    : 92.388942 seconds
  Total time spent in hash processes                       : 19.472116 seconds
  Total time spent in compute processes                    : 72.916826 seconds
  Total time spent per hash process                        : 0.000008 seconds
  Total time spent per compute process                     : 0.000237 seconds
  Total number of hash processes                           : 2351184
  Total number of compute processes                        : 307995
  Total number of hash processes per compute process       : 7.633838
```

Here there is definitely a small improvement - I haven't narrowed down which fixes had a positive effect in this case, or whether any change in isolation could have had a small negative impact, but the end result is not bad at least.

What I really like about these changes overall is that they make weirdness easier to spot.  Running with all the fixes on, I took a look at the performance monitor breakdown, and there was one expression that was being evaluated per tile and shouldn't have been.  It turns out that a user was using a generic SwitchComputeNode with image plugs instead of an ImageSwitch, so the index wasn't getting the context trimmed.  ( Any ideas how to discourage this? )